+++ persistence/src/main/java/com/example/model/SongCommentReaction.java
@@ -36,7 +36,7 @@ public class SongCommentReaction {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private Integer value;
+    private short value;
 
     private LocalDateTime createdAt;
